requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
alabaster==0.7.12
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asttokens==2.0.5
attrs==21.4.0
awscli==1.22.83
Babel==2.9.1
backcall==0.2.0
binaryornot==0.4.4
black==21.12b0
bleach==4.1.0
blinker==1.4
botocore==1.24.28
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.3
coloredlogs==15.0.1
cookiecutter==1.7.3
coverage==6.3.2
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docutils==0.15.2
dtaidistance==2.3.6
entrypoints==0.4
executing==0.8.3
flake8==4.0.1
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.27
humanfriendly==10.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
ipykernel==6.9.1
ipympl==0.8.8
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy==0.931
mypy-extensions==0.4.3
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
pmdarima==1.8.5
poyo==0.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
pyflakes==2.4.0
pygame==2.1.2
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pyreadline3==3.3
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
rsa==4.7.2
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
-e git+https://github.com/spicytelescope/INSA_4A_STI_TD.git@95162db191de7906c220b5b2f724c9e9a9e6bdce#egg=src&subdirectory=00_ia_avancee\05_covid_project_ia
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.8.1
tabulate==0.8.9
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
